requirements.txt
appdirs==1.4.4
asgiref==3.4.1
distlib==0.3.2
Django==3.2.5
filelock==3.0.12
pytz==2021.1
gunicorn
six==1.16.0
whitenoise
sqlparse==0.2.4
virtualenv==20.4.7
python-decouple
dnspython==2.1.0
pymongo==3.12.0
djongo
Pillow==8.3.1

